requirements.txt
python-telegram-bot==20.7
flask==2.3.3
python-dotenv==1.0.0
reportlab==4.0.7
gunicorn==21.2.0
pytz==2023.3
quart==0.17.6